Refactoring common code out into common file
