+++ dask_mpi/cli.py
@@ -1,15 +1,10 @@
-from functools import partial
-
 import click
 from mpi4py import MPI
 from tornado.ioloop import IOLoop
-from tornado import gen
 
-from distributed import Scheduler, Nanny, Worker
-from distributed.bokeh.worker import BokehWorker
-from distributed.cli.utils import check_python_3, uri_from_host_port
-from distributed.utils import get_ip_interface
+from distributed.cli.utils import check_python_3
 
+from dask_mpi.common import get_host_from_interface, start_scheduler, start_worker
 
 comm = MPI.COMM_WORLD
 rank = comm.Get_rank()
@@ -43,57 +38,14 @@ loop = IOLoop()
               help="Prefix for the bokeh app")
 def main(scheduler_file, interface, nthreads, local_directory, memory_limit,
          scheduler, bokeh_port, bokeh_prefix, nanny, bokeh_worker_port):
-    if interface:
-        host = get_ip_interface(interface)
-    else:
-        host = None
+    host = get_host_from_interface(interface)
 
     if rank == 0 and scheduler:
-        try:
-            from distributed.bokeh.scheduler import BokehScheduler
-        except ImportError:
-            services = {}
-        else:
-            services = {('bokeh',  bokeh_port): partial(BokehScheduler,
-                                                        prefix=bokeh_prefix)}
-        scheduler = Scheduler(scheduler_file=scheduler_file,
-                              loop=loop,
-                              services=services)
-        addr = uri_from_host_port(host, None, 8786)
-        scheduler.start(addr)
-        try:
-            loop.start()
-            loop.close()
-        finally:
-            scheduler.stop()
+        start_scheduler(host, loop, scheduler_file, bokeh_port, bokeh_prefix)
     else:
-        W = Nanny if nanny else Worker
-        worker = W(scheduler_file=scheduler_file,
-                   loop=loop,
-                   name=rank if scheduler else None,
-                   ncores=nthreads,
-                   local_dir=local_directory,
-                   services={('bokeh', bokeh_worker_port): BokehWorker},
-                   memory_limit=memory_limit)
-        addr = uri_from_host_port(host, None, 0)
-
-        @gen.coroutine
-        def run():
-            yield worker._start(addr)
-            while worker.status != 'closed':
-                yield gen.sleep(0.2)
-
-        try:
-            loop.run_sync(run)
-            loop.close()
-        finally:
-            pass
-
-        @gen.coroutine
-        def close():
-            yield worker._close(timeout=2)
-
-        loop.run_sync(close)
+        name = rank if scheduler else None
+        start_worker(host, loop, name, scheduler_file, nanny, local_directory,
+                     nthreads, memory_limit, bokeh_worker_port)
 
 
 def go():
